Fix: deleted DAO, moved file
