+++ project/src/test/java/com/favorite/project/User/LoginServiceTest.java
@@ -1,4 +1,4 @@
-package com.favorite.project;
+package com.favorite.project.User;
 
 
 import com.favorite.project.User.dto.LoginForm;
